+++ src/Ozds.Data/Migrations/20260205090933_AddCorrectIndexForAuditEvent.cs
@@ -1,0 +1,27 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Ozds.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddCorrectIndexForAuditEvent : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateIndex(
+                name: "ix_events_auditable_entity_id",
+                table: "events",
+                column: "auditable_entity_id");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropIndex(
+                name: "ix_events_auditable_entity_id",
+                table: "events");
+        }
+    }
+}